requirements.txt
flet==0.18.0
requests==2.31.0
pillow==10.0.0
numpy==1.24.3
opencv-python==4.12.0.88